Extract `PciDeviceLocation` in ostd into PCI component
