+++ kernel/comps/pci/src/common_device.rs
@@ -6,13 +6,12 @@
 
 use alloc::vec::Vec;
 
-use ostd::bus::pci::PciDeviceLocation;
-
 use super::{
     capability::Capability,
     cfg_space::{AddrLen, Bar, Command, PciDeviceCommonCfgOffset, Status},
     device_info::PciDeviceId,
 };
+use crate::device_info::PciDeviceLocation;
 
 /// PCI common device, Contains a range of information and functions common to PCI devices.
 #[derive(Debug)]
